+++ src/monorepo_npm_check_updates_cli.ts
@@ -1,13 +1,13 @@
-import { _param_h } from '@ctx-core/cli-args'
+import { param_r_ } from '@ctx-core/cli-args'
 import { monorepo_npm_check_updates } from './monorepo_npm_check_updates'
 export async function monorepo_npm_check_updates_cli() {
-	const param_h = _param_h(process.argv.slice(2), {
+	const param_r = param_r_(process.argv.slice(2), {
 		threads: '-t, --threads',
 		workspace_name: '-w, --workspace-name'
 	}, {
 		threads: 20,
 	})
-	const stdout_h_package_name = await monorepo_npm_check_updates(param_h)
+	const stdout_h_package_name = await monorepo_npm_check_updates(param_r)
 	for (let package_name in stdout_h_package_name) {
 		const stdout = stdout_h_package_name[package_name]
 		if (stdout) {