+++ src/run_parallel_workspaces.ts
@@ -1,4 +1,4 @@
-import { _queue } from '@ctx-core/queue'
+import { queue_ } from '@ctx-core/queue'
 import { projects_ } from './projects_'
 import { promise_a_ } from './promise_a_'
 import { exec } from './exec'
@@ -6,7 +6,7 @@ import type { monorepo_thread_opts_T } from './monorepo_thread_opts_T'
 import type { project_T } from './project_T'
 import { stdout_h_package_name_ } from './stdout_h_package_name_'
 export async function run_parallel_workspaces(cmd_a:string[], opts:monorepo_thread_opts_T = {}) {
-	const queue = _queue(opts.threads || 20)
+	const queue = queue_(opts.threads || 20)
 	const projects = await projects_()
 	const cmd = cmd_a.join(' ')
 	const package_name_a = projects.map(project=>project.package_name)