+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ctx-core/monorepo
 
+## 17.2.9
+
+### Patch Changes
+
+- use param*r* instead of \_param_r
+- use queue\_ instead of \_queue
+- Updated dependencies [undefined]
+  - @ctx-core/cli-args@9.0.17
+
 ## 17.2.8
 
 ### Patch Changes